+++ packages/@office-iss/react-native-win32/src/Libraries/Image/Image.win32.js
@@ -148,7 +148,9 @@ const BaseImage = (props: ImagePropsType, forwardedRef) => {
     }
   }
 
+  // $FlowFixMe[prop-missing]
   const resizeMode = props.resizeMode || style.resizeMode || 'cover';
+  // $FlowFixMe[prop-missing]
   const tintColor = style.tintColor;
 
   if (props.src != null) {