+++ packages/@react-native-windows/codegen/src/generators/ParamTypes.ts
@@ -6,7 +6,7 @@
 
 'use strict';
 
-import {
+import type {
   NamedShape,
   NativeModuleParamTypeAnnotation,
   Nullable,