+++ packages/@react-native-windows/automation-channel/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @react-native-windows/automation-channel
 
-This log was last generated on Fri, 30 Sep 2022 05:11:35 GMT and should not be manually modified.
+This log was last generated on Sat, 01 Oct 2022 05:11:34 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.132
+
+Sat, 01 Oct 2022 05:11:34 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.558
 
 ## 0.1.131